+++ presentation/src/main/java/com/mashup/presentation/detail/chat/compose/ChatDetailScreen.kt
@@ -1,21 +1,21 @@
 package com.mashup.presentation.detail.chat.compose
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.mashup.presentation.R
 import com.mashup.presentation.ui.common.KeyLinkToolbar
-import com.mashup.presentation.ui.theme.*
+import com.mashup.presentation.ui.theme.Black
+import com.mashup.presentation.ui.theme.SsamDTheme
+import com.mashup.presentation.ui.theme.White
 
 /**
  * Ssam_D_Android
@@ -58,50 +58,17 @@ fun ChatDetailContent(
         OtherUserInfo(
             modifier = Modifier.padding(top = 8.dp)
         )
-    }
-}
 
-@Composable
-fun OtherUserInfo(
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 85.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Box(
+        MatchedKeywords(
             modifier = Modifier
-                .size(48.dp)
-                .background(
-                    color = Red,
-                    shape = CircleShape
-                ),
-        )
-
-        Text(
-            text = "슈퍼 니카2",
-            style = Title2,
-            color = White,
-            modifier = Modifier.padding(top = 8.dp)
         )
     }
 }
-
 
 @Preview(showBackground = true)
 @Composable
 private fun ChatDetailScreenPreview() {
     SsamDTheme {
         ChatDetailScreen()
-    }
-}
-
-@Preview(showBackground = true, backgroundColor = 0XFF0A0C10)
-@Composable
-private fun OtherUserInfoPreview() {
-    SsamDTheme {
-        OtherUserInfo()
     }
 }
