+++ presentation/src/main/java/com/mashup/presentation/detail/chat/compose/MatchedKeywords.kt
@@ -1,0 +1,15 @@
+package com.mashup.presentation.detail.chat.compose
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+/**
+ * Ssam_D_Android
+ * @author jaesung
+ * @created 2023/06/28
+ */
+
+@Composable
+fun MatchedKeywords(modifier: Modifier = Modifier) {
+
+}
