+++ presentation/src/main/java/com/mashup/presentation/detail/chat/compose/OtherUserInfo.kt
@@ -1,0 +1,57 @@
+package com.mashup.presentation.detail.chat.compose
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.mashup.presentation.ui.theme.Red
+import com.mashup.presentation.ui.theme.SsamDTheme
+import com.mashup.presentation.ui.theme.Title2
+import com.mashup.presentation.ui.theme.White
+
+/**
+ * Ssam_D_Android
+ * @author jaesung
+ * @created 2023/06/28
+ */
+
+@Composable
+fun OtherUserInfo(
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(horizontal = 85.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Box(
+            modifier = Modifier
+                .size(48.dp)
+                .background(
+                    color = Red,
+                    shape = CircleShape
+                ),
+        )
+
+        Text(
+            text = "슈퍼 니카2",
+            style = Title2,
+            color = White,
+            modifier = Modifier.padding(top = 8.dp)
+        )
+    }
+}
+
+@Preview(showBackground = true, backgroundColor = 0XFF0A0C10)
+@Composable
+private fun OtherUserInfoPreview() {
+    SsamDTheme {
+        OtherUserInfo()
+    }
+}
